pyproject.toml
[project]
name = "llm-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.21",
    "arxiv>=2.3.1",
    "fastapi>=0.128.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "openai>=2.14.0",
    "praw>=7.8.1",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
]
